+++ tensorboard/data/server/lib.rs
@@ -15,12 +15,15 @@ limitations under the License.
 
 //! Core functionality for TensorBoard data loading.
 
+#![allow(clippy::needless_update)] // https://github.com/rust-lang/rust-clippy/issues/6323
+
 pub mod commit;
 pub mod data_compat;
 pub mod event_file;
 pub mod masked_crc;
 pub mod reservoir;
 pub mod run;
+pub mod server;
 pub mod tf_record;
 pub mod types;
